+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Extract Variable won't extract type parameter instantiations anymore, because it shouldn't!
+- Split Declaration from Initialization now works fine with destructured assignments.
 
 ## [5.2.0] - 2021-04-15 - Poor Unfortunate Types 🦑